Psuedo Backboard TO DO attached.
Wrote the pseudo code for the backboard scoring, gotta work out kinks when I actually get my hands on the robot but this'll make it easier. Figured only do it on side closest to backboard or we have to worry about interference with alliance.
TO-DO:
in the TensorCore file change tolerances to what they would be now with new camera and line up placement.
Either do backboard on side closest to backboard, spike mark on side furtherest, or combine. This'll take a ton of time so please finish soon, please.