Autonomous Red Side fully working (Kinda)
The red side of autonomous is completely functioning with a tad bit of hiccups but nothing to really worry about.
Next is blue side which is just red but mirrored which I'll HAVE to math out because we don't have a blue side backboard.
Miss Bauer also wants me to add a version that just does spike and park because 5 points is pretty dang helpful.